10 add new block to db periodically

+++ models/block.go
@@ -6,7 +6,7 @@ import "time"
 type Block struct {
 	ID             int64     `db:"id"`
 	Hash           string    `db:"hash"`
-	Height         int32     `db:"height"`
+	Height         int64     `db:"height"`
 	BlockCreatedAt time.Time `db:"block_created_at"`
 	CreatedAt      time.Time `db:"created_at"`
 }
